refactor(step3): 패키지 명 변경
