+++ src/main/kotlin/calculator/tasks/Calculator.kt
@@ -1,0 +1,28 @@
+package calculator.tasks
+
+import calculator.data.Operand
+import calculator.data.Operator
+import calculator.util.extension.toDouble
+import calculator.util.Const.Companion.DELIMITERS_BLANK
+import calculator.util.Const.Companion.NOTICE_NOT_OPERAND_SYMBOL
+import calculator.util.extension.splitToCalculate
+
+class Calculator {
+    fun splitReadLine(): List<String> {
+        return readLine().splitToCalculate(DELIMITERS_BLANK)
+    }
+
+    fun calculate(anyOfCollection: List<String>): String {
+        return anyOfCollection.reduceIndexed { index, acc, element ->
+            if (index % 2 != 1 && index > 0) {
+                val operator =
+                    Operator.findByOperatorSymbol(anyOfCollection[index - 1])
+                return@reduceIndexed operator.calculate(
+                    Operand(acc.toDouble(NOTICE_NOT_OPERAND_SYMBOL)),
+                    Operand(element.toDouble(NOTICE_NOT_OPERAND_SYMBOL))
+                ).toString()
+            }
+            return@reduceIndexed acc
+        }
+    }
+}
