+++ src/main/kotlin/racingcar/tasks/CarFactory.kt
@@ -1,0 +1,7 @@
+package racingcar.tasks
+
+import racingcar.model.Car
+
+class CarFactory {
+    fun createCar() = Car()
+}
