+++ src/main/kotlin/racingcar/ui/OutputView.kt
@@ -1,0 +1,7 @@
+package racingcar.ui
+
+import racingcar.model.Car
+
+fun getResult(car: Car) {
+    println(car.whereIs())
+}
