+++ src/main/kotlin/racingcar/util/Message.kt
@@ -1,0 +1,13 @@
+package racingcar.util
+
+@Suppress("SpellCheckingInspection")
+class Message {
+    companion object {
+        const val NOTICE_INPUT_TRY_COUNT = "시도횟수를 입력해주세요."
+        const val NOTICE_INPUT_CAR_COUNT = "자동차 대수는 몇 대인가요?"
+        const val ONE_DISTANCE = "-"
+        const val NOTICE_CANNOT_CAST_TO_INT = "을(를) Int형으로 변환할 수 없습니다."
+        const val MAX_CONDITION = 10
+        const val MIN_CONDITION_FOR_GOING = 4
+    }
+}
