--- src/main/kotlin/step3/tasks/CarFactory.kt
@@ -1,7 +1,0 @@
-package step3.tasks
-
-import step3.model.Car
-
-class CarFactory {
-    fun createCar() = Car()
-}
