--- src/main/kotlin/step3/ui/OutputView.kt
@@ -1,7 +1,0 @@
-package step3.ui
-
-import step3.model.Car
-
-fun getResult(car: Car) {
-    println(car.whereIs())
-}
