+++ src/test/kotlin/racingcar/tasks/CarTest.kt
@@ -1,0 +1,21 @@
+package racingcar.tasks
+
+import org.assertj.core.api.Assertions
+import org.junit.jupiter.api.Test
+
+import racingcar.model.Car
+import racingcar.util.Message.Companion.ONE_DISTANCE
+
+@Suppress("NonAsciiCharacters")
+class CarTest {
+
+    @Test
+    fun `자동차 1번 전진`() {
+        val car = Car().apply {
+            go()
+        }
+
+        Assertions.assertThat(car.whereIs())
+            .isEqualTo(ONE_DISTANCE)
+    }
+}
